Moves state change filtering to the background script, next step is adding this to the videoHistory.
